initial structure implemented
