+++ src/index.ts
@@ -24,4 +24,10 @@ export {
 	initKeypom,
 	getEnv,
 } from "./lib/keypom";
-export * from "./lib/types";
+export * from "./lib/types/drops";
+export * from "./lib/types/fc";
+export * from "./lib/types/ft";
+export * from "./lib/types/general";
+export * from "./lib/types/nft";
+export * from "./lib/types/params";
+export * from "./lib/types/simple";
