+++ src/lib/types/fc.ts
@@ -1,0 +1,12 @@
+export interface Method {
+	receiverId: string;
+	methodName: string;
+	args: string;
+	attachedDeposit: string;
+	accountIdField: string;
+	dropIdField: string;
+}
+
+export interface FCData {
+	methods: Method[][]
+}
