+++ src/lib/types/ft.ts
@@ -1,0 +1,5 @@
+export interface FTData {
+	contractId?: string;
+	senderId?: string;
+	balancePerUse?: string;
+}
