+++ src/lib/types/nft.ts
@@ -1,0 +1,5 @@
+export interface NFTData {
+	contractId?: string;
+	senderId?: string;
+	tokenIds?: string[];
+}
