+++ src/lib/types/simple.ts
@@ -1,0 +1,4 @@
+export interface SimpleData {
+	// If this is set to true, keys can be created and registered AFTER they've been created (for simple and FC drops only).
+    lazyRegister?: boolean,
+}
